requirements.txt
# CineMatch V2.1.6 - Python Dependencies
# Security: Pinned versions verified with pip-audit (Dec 2025)

# ========================================
# Core Data Science & ML Libraries
# ========================================
pandas==2.3.3                       # Pinned for security
numpy==1.26.4                       # NumPy 1.x required for scikit-surprise compatibility
scikit-learn==1.7.2                 # Python 3.13 compatible
scipy==1.15.3                       # Scientific computing
scikit-surprise==1.1.4              # Collaborative filtering library for SVD
pyarrow==18.1.0                     # Parquet support for 5x faster data loading

# ========================================
# Web Framework & UI
# ========================================
streamlit==1.51.0                   # Interactive web interface (pinned)

# ========================================
# Data Visualization
# ========================================
plotly==6.5.0                       # Interactive charts
matplotlib==3.10.7                  # Static plots  
seaborn==0.13.2                     # Statistical visualizations

# ========================================
# Model Serialization & Persistence
# ========================================
joblib==1.3.2                       # Model saving/loading

# ========================================
# Utilities
# ========================================
python-dotenv==1.0.0                # Environment variable management

# ========================================
# Security Dependencies
# ========================================
# Note: pip-audit used for vulnerability scanning (dev only)
